requirements.txt
openclaw>=0.9.0
playwright>=1.40.0
requests>=2.31.0
pyyaml>=6.0.1
schedule>=1.2.0
python-dateutil>=2.8.2
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
fake-useragent>=1.4.0